[#57] Feat: Multipart 모델 변경
- PresignedURL에 맞춤변경
