+++ Projects/Core/HGNetwork/Sources/Protocol/Multipart/MimeType.swift
@@ -1,0 +1,13 @@
+//
+//  MimeType.swift
+//  HGNetwork
+//
+//  Created by Enes on 7/5/25.
+//
+
+import Foundation
+
+public enum MimeType: String {
+    case jpeg = "image/jpeg"
+    case png = "image/png"
+}
