+++ Projects/Core/HGNetwork/Sources/Types/BaseURL.swift
@@ -9,5 +9,4 @@ import Foundation
 
 public enum BaseURL: String {
     case host = "hgdgds.duckdns.org"
-    case image = "my-app-image-bucket.s3.ap-northeast-2.amazonaws.com"
 }
